Addressed PR suggestions
