+++ lib/telemetry_metrics_statsd/emitter.ex
@@ -1,5 +1,5 @@
 defmodule TelemetryMetricsStatsd.Emitter do
   @moduledoc false
   @callback emit(name :: GenServer.name(), metric :: iodata()) :: :ok
-  @callback emit_internal(name :: GenServer.name(), metric :: iodata) :: :ok
+  @callback emit_internal(name :: GenServer.name(), metric :: iodata()) :: :ok
 end
