+++ lib/telemetry_metrics_statsd/event_handler.ex
@@ -4,6 +4,15 @@ defmodule TelemetryMetricsStatsd.EventHandler do
   alias Telemetry.Metrics
   alias TelemetryMetricsStatsd.Formatter
 
+  alias TelemetryMetricsStatsd.Options
+  use GenServer
+
+  # Public
+
+  def start_link([options, emitter_module]) do
+    GenServer.start_link(__MODULE__, [options, emitter_module])
+  end
+
   @spec attach(
           GenServer.name(),
           [Metrics.t()],
@@ -68,6 +77,36 @@ defmodule TelemetryMetricsStatsd.EventHandler do
       publish_metrics(emitter_module, name, metrics)
     end
   end
+
+  # OTP
+  @impl true
+  def init([%Options{} = options, emitter_module]) do
+    Process.flag(:trap_exit, true)
+
+    handler_ids =
+      attach(
+        options.name,
+        options.metrics,
+        emitter_module,
+        options.prefix,
+        options.formatter,
+        options.global_tags
+      )
+
+    {:ok, handler_ids}
+  end
+
+  @impl true
+  def handle_info({:EXIT, _pid, reason}, handler_ids) do
+    {:stop, reason, handler_ids}
+  end
+
+  @impl true
+  def terminate(_reason, handler_ids) do
+    detach(handler_ids)
+  end
+
+  # Private
 
   @spec handler_id(
           registered_name :: GenServer.name(),
